Move reader node definition together with kind
