+++ dds/src/dds/domain/dcps_service.rs
@@ -30,8 +30,7 @@ use crate::{
                 ENTITYID_SEDP_BUILTIN_TOPICS_ANNOUNCER, ENTITYID_SEDP_BUILTIN_TOPICS_DETECTOR,
             },
             dds_subscriber::DdsSubscriber,
-            node_kind::SubscriberNodeKind,
-            node_user_defined_data_reader::UserDefinedDataReaderNode,
+            node_kind::{SubscriberNodeKind, UserDefinedDataReaderNode},
             node_user_defined_data_writer::UserDefinedDataWriterNode,
             node_user_defined_subscriber::UserDefinedSubscriberNode,
             node_user_defined_topic::UserDefinedTopicNode,
