+++ dds/src/implementation/dds_impl/node_kind.rs
@@ -1,6 +1,7 @@
+use crate::{implementation::rtps::types::Guid, subscription::data_reader::AnyDataReader};
+
 use super::{
     node_builtin_subscriber::BuiltinSubscriberNode,
-    node_user_defined_data_reader::UserDefinedDataReaderNode,
     node_user_defined_data_writer::UserDefinedDataWriterNode,
     node_user_defined_subscriber::UserDefinedSubscriberNode,
     node_user_defined_topic::UserDefinedTopicNode,
@@ -32,3 +33,34 @@ pub enum TopicNodeKind {
     UserDefined(UserDefinedTopicNode),
     Listener(UserDefinedTopicNode),
 }
+
+#[derive(PartialEq, Eq, Debug, Clone, Copy)]
+pub struct UserDefinedDataReaderNode {
+    this: Guid,
+    parent_subcriber: Guid,
+    parent_participant: Guid,
+}
+
+impl UserDefinedDataReaderNode {
+    pub fn new(this: Guid, parent_subcriber: Guid, parent_participant: Guid) -> Self {
+        Self {
+            this,
+            parent_subcriber,
+            parent_participant,
+        }
+    }
+
+    pub fn guid(&self) -> Guid {
+        self.this
+    }
+
+    pub fn parent_subscriber(&self) -> Guid {
+        self.parent_subcriber
+    }
+
+    pub fn parent_participant(&self) -> Guid {
+        self.parent_participant
+    }
+}
+
+impl AnyDataReader for UserDefinedDataReaderNode {}
