+++ dds/src/implementation/dds_impl/node_user_defined_data_reader.rs
@@ -11,7 +11,7 @@ use crate::{
         },
     },
     subscription::{
-        data_reader::{AnyDataReader, Sample},
+        data_reader::Sample,
         sample_info::{InstanceStateKind, SampleStateKind, ViewStateKind},
     },
     topic_definition::type_support::DdsDeserialize,
@@ -22,37 +22,6 @@ use super::{
     node_user_defined_subscriber::UserDefinedSubscriberNode,
     node_user_defined_topic::UserDefinedTopicNode,
 };
-
-#[derive(PartialEq, Eq, Debug, Clone, Copy)]
-pub struct UserDefinedDataReaderNode {
-    this: Guid,
-    parent_subcriber: Guid,
-    parent_participant: Guid,
-}
-
-impl UserDefinedDataReaderNode {
-    pub fn new(this: Guid, parent_subcriber: Guid, parent_participant: Guid) -> Self {
-        Self {
-            this,
-            parent_subcriber,
-            parent_participant,
-        }
-    }
-
-    pub fn guid(&self) -> Guid {
-        self.this
-    }
-
-    pub fn parent_subscriber(&self) -> Guid {
-        self.parent_subcriber
-    }
-
-    pub fn parent_participant(&self) -> Guid {
-        self.parent_participant
-    }
-}
-
-impl AnyDataReader for UserDefinedDataReaderNode {}
 
 pub fn read<Foo>(
     domain_participant: &mut DdsDomainParticipant,
