+++ dds/src/implementation/dds_impl/status_listener.rs
@@ -4,9 +4,8 @@ use crate::{
 };
 
 use super::{
-    node_user_defined_data_reader::UserDefinedDataReaderNode,
-    node_user_defined_data_writer::UserDefinedDataWriterNode,
-    status_condition_impl::StatusConditionImpl, node_user_defined_topic::UserDefinedTopicNode,
+    node_kind::UserDefinedDataReaderNode, node_user_defined_data_writer::UserDefinedDataWriterNode,
+    node_user_defined_topic::UserDefinedTopicNode, status_condition_impl::StatusConditionImpl,
 };
 
 pub struct StatusListener<T: ?Sized> {
